Added feature to plot Labels below figure
removed entire sequence of numbers from intermediate mass elements and
moved labels to bottom of plot.  Also updated with code to save legends
with the plot.  Exploring N for frequency spectrum.  Larger N's are
actually causing enormous peak amplitudes over a short duration followed
by very low amplitude for a long duration, then repeats.  This behavior
is also seen in the ship and string mass positions/velocities.  Created
Plot for Wave Amplitude as a function of frequency to understand the
shape of the spectrum for different values of N frequency divisions.